requirement.txt
python-3.10.12
streamlit
git+https://github.com/pymupdf/PyMuPDF.git
pandas==2.2.1
git+https://github.com/pymupdf/PyMuPDF.git@master
llama-index
pdfplumber
chromadb
google-genai==1.26.0
google.api_core
pytesseract
transformers
dotenv







